+++ infra/resources/prod/azure.tf
@@ -1,6 +1,6 @@
 module "azure_core_values" {
   source  = "pagopa-dx/azure-core-values-exporter/azurerm"
-  version = "~> 0.0"
+  version = "~> 0.4"
 
   core_state = local.core_state
 }